+++ internal/command/jsonstate/state_test.go
@@ -1054,3 +1054,210 @@ func TestSensitiveAsBool(t *testing.T) {
 		}
 	}
 }
+
+func TestSensitiveAsBoolWithPathValueMarks(t *testing.T) {
+	tests := []struct {
+		Input cty.Value
+		Pvms  []cty.PathValueMarks
+		Want  cty.Value
+	}{
+		{
+			cty.ListVal([]cty.Value{
+				cty.StringVal("hello"),
+				cty.StringVal("friend"),
+			}),
+			[]cty.PathValueMarks{{
+				Path:  cty.Path{cty.IndexStep{Key: cty.NumberIntVal(1)}},
+				Marks: cty.NewValueMarks(marks.Sensitive)},
+			},
+			cty.TupleVal([]cty.Value{
+				cty.False,
+				cty.True,
+			}),
+		},
+		{
+			cty.ListVal([]cty.Value{
+				cty.StringVal("hello").Mark(marks.Sensitive),
+				cty.StringVal("friend"),
+			}),
+			[]cty.PathValueMarks{{
+				Path:  cty.Path{cty.IndexStep{Key: cty.NumberIntVal(1)}},
+				Marks: cty.NewValueMarks(marks.Sensitive)},
+			},
+			cty.TupleVal([]cty.Value{
+				cty.True,
+				cty.True,
+			}),
+		},
+		{
+			cty.TupleVal([]cty.Value{
+				cty.StringVal("hello"),
+				cty.StringVal("friend"),
+			}),
+			[]cty.PathValueMarks{{
+				Path:  cty.Path{cty.IndexStep{Key: cty.NumberIntVal(1)}},
+				Marks: cty.NewValueMarks(marks.Sensitive)},
+			},
+			cty.TupleVal([]cty.Value{
+				cty.False,
+				cty.True,
+			}),
+		},
+		{
+			cty.TupleVal([]cty.Value{
+				cty.StringVal("hello").Mark(marks.Sensitive),
+				cty.StringVal("friend"),
+			}),
+			[]cty.PathValueMarks{{
+				Path:  cty.Path{cty.IndexStep{Key: cty.NumberIntVal(1)}},
+				Marks: cty.NewValueMarks(marks.Sensitive)},
+			},
+			cty.TupleVal([]cty.Value{
+				cty.True,
+				cty.True,
+			}),
+		},
+		{
+			cty.SetVal([]cty.Value{
+				cty.StringVal("hello"),
+				cty.StringVal("friend"),
+			}),
+			[]cty.PathValueMarks{{
+				Path:  cty.Path{cty.IndexStep{Key: cty.StringVal("hello")}},
+				Marks: cty.NewValueMarks(marks.Sensitive)},
+			},
+			cty.TupleVal([]cty.Value{
+				cty.False,
+				cty.True,
+			}),
+		},
+		{
+			cty.MapVal(map[string]cty.Value{
+				"greeting": cty.StringVal("hello"),
+				"animal":   cty.StringVal("horse"),
+			}),
+			[]cty.PathValueMarks{{
+				Path:  cty.Path{cty.IndexStep{Key: cty.StringVal("animal")}},
+				Marks: cty.NewValueMarks(marks.Sensitive)},
+			},
+			cty.ObjectVal(map[string]cty.Value{
+				"animal": cty.True,
+			}),
+		},
+		{
+			cty.ObjectVal(map[string]cty.Value{
+				"greeting": cty.StringVal("hello"),
+				"animal":   cty.StringVal("horse"),
+			}),
+			[]cty.PathValueMarks{{
+				Path:  cty.Path{cty.GetAttrStep{Name: "animal"}},
+				Marks: cty.NewValueMarks(marks.Sensitive)},
+			},
+			cty.ObjectVal(map[string]cty.Value{
+				"animal": cty.True,
+			}),
+		},
+		{
+			cty.ListVal([]cty.Value{
+				cty.ObjectVal(map[string]cty.Value{
+					"a": cty.UnknownVal(cty.String),
+				}),
+				cty.ObjectVal(map[string]cty.Value{
+					"a": cty.StringVal("known"),
+				}),
+			}),
+			[]cty.PathValueMarks{{
+				Path:  cty.Path{cty.IndexStep{Key: cty.NumberIntVal(1)}, cty.GetAttrStep{Name: "a"}},
+				Marks: cty.NewValueMarks(marks.Sensitive)},
+			},
+
+			cty.TupleVal([]cty.Value{
+				cty.EmptyObjectVal,
+				cty.ObjectVal(map[string]cty.Value{
+					"a": cty.True,
+				}),
+			}),
+		},
+		{
+			cty.TupleVal([]cty.Value{
+				cty.ObjectVal(map[string]cty.Value{
+					"a": cty.UnknownVal(cty.String),
+				}),
+				cty.ObjectVal(map[string]cty.Value{
+					"a": cty.StringVal("known"),
+				}),
+			}),
+			[]cty.PathValueMarks{{
+				Path:  cty.Path{cty.IndexStep{Key: cty.NumberIntVal(1)}, cty.GetAttrStep{Name: "a"}},
+				Marks: cty.NewValueMarks(marks.Sensitive)},
+			},
+
+			cty.TupleVal([]cty.Value{
+				cty.EmptyObjectVal,
+				cty.ObjectVal(map[string]cty.Value{
+					"a": cty.True,
+				}),
+			}),
+		},
+		{
+			cty.ListVal([]cty.Value{
+				cty.MapValEmpty(cty.String),
+				cty.MapVal(map[string]cty.Value{
+					"a": cty.StringVal("known"),
+				}),
+				cty.MapVal(map[string]cty.Value{
+					"a": cty.UnknownVal(cty.String),
+				}),
+			}),
+			[]cty.PathValueMarks{{
+				Path:  cty.Path{cty.IndexStep{Key: cty.NumberIntVal(1)}, cty.IndexStep{Key: cty.StringVal("a")}},
+				Marks: cty.NewValueMarks(marks.Sensitive)},
+			},
+			cty.TupleVal([]cty.Value{
+				cty.EmptyObjectVal,
+				cty.ObjectVal(map[string]cty.Value{
+					"a": cty.True,
+				}),
+				cty.EmptyObjectVal,
+			}),
+		},
+		{
+			cty.SetVal([]cty.Value{
+				cty.ObjectVal(map[string]cty.Value{
+					"greeting": cty.StringVal("hello"),
+					"animal":   cty.StringVal("cat"),
+				}),
+				cty.ObjectVal(map[string]cty.Value{
+					"greeting": cty.StringVal("hello"),
+					"animal":   cty.StringVal("horse"),
+				}),
+			}),
+			[]cty.PathValueMarks{{
+				Path: cty.Path{
+					cty.IndexStep{Key: cty.ObjectVal(map[string]cty.Value{
+						"greeting": cty.StringVal("hello"),
+						"animal":   cty.StringVal("cat"),
+					})},
+					cty.GetAttrStep{Name: "animal"}},
+				Marks: cty.NewValueMarks(marks.Sensitive)},
+			},
+
+			cty.TupleVal([]cty.Value{
+				cty.ObjectVal(map[string]cty.Value{
+					"animal": cty.True,
+				}),
+				cty.EmptyObjectVal,
+			}),
+		},
+	}
+
+	for _, test := range tests {
+		got := SensitiveAsBoolWithPathValueMarks(test.Input, test.Pvms)
+		if !reflect.DeepEqual(got, test.Want) {
+			t.Errorf(
+				"wrong result\ninput: %#v\ngot:   %#v\nwant:  %#v",
+				test.Input, got, test.Want,
+			)
+		}
+	}
+}